fix: button and location fields are now components
